cdk.out/asset.352e51a9de5e2b073e96737d326518c983014dd91f51dd732ef63c262db0763c/pyproject.toml
[project]
name = "fomc-agent"
version = "0.1.0"
description = "FOMC data pipeline: BLS & DataUSA ingestion, analytics, and serverless automation"
requires-python = ">=3.12"
dependencies = [
    "boto3",
    "pandas",
]

[project.optional-dependencies]
cdk = ["aws-cdk-lib", "constructs"]
dev = [
    "pytest",
    "pytest-cov",
    "moto[s3,sqs,lambda,iam]",
    "freezegun",
    "matplotlib",
]
